+++ CHANGELOG.md
@@ -3,6 +3,12 @@ FlatLaf Change Log
 
 ## 0.44-SNAPSHOT
 
+#### New features and improvements
+
+- Support painting separator line between window title and content (use UI value
+  `TitlePane.borderColor`). (issue #184)
+
+
 #### Fixed bugs
 
 - Custom window decorations: Not visible menu bar is now ignored in layout.